requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
tweepy
Pillow
pytesseract
python-multipart